requirements.txt
cryptography
pyjwt==1.5.3
pylint
requests
